XIVY-2942 refactor: extract data class detail components
